Refactor json_response body
Modern versions of FireFox the JSON content for sessionId, capabilities, etc is buried in the enclosing "value" tag. 
This is going to presume it's a modern version and strip that level to the inside contents of a TCL dictionary. 
We further use that json_response for 'capabilities'
Further use that to test for W3C compliance as well